*extend timestamp counter to 48bit

+++ Src/xcan_timestamp.h
@@ -3,7 +3,7 @@
 
 void xcan_timestamp_init( void );
 uint32_t xcan_timestamp_millis( void );
-uint32_t xcan_timestamp_us( void );
+uint64_t xcan_timestamp_us( void );
 void xcan_timestamp_ticks( uint16_t *ptime );
-void xcan_timestamp_ticks_from_ts( uint16_t *ptime, uint32_t ts );
+void xcan_timestamp_ticks_from_ts( uint16_t *ptime, uint64_t ts );
